+++ src/test_runner.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <algorithm>
 #include <cstddef>
 #include <string>
 #include <string_view>
@@ -114,6 +115,16 @@ void update_scenario_status(std::string_view name, std::string_view file,
   {
     results::scenarios_back().status = results::test_status::skipped;
   }
+  else if (internal::get_runtime_options().fail_scenario().is_set)
+  {
+    const std::string& msg =
+        internal::get_runtime_options().fail_scenario().msg;
+    log::error(msg, log::new_line);
+    results::scenarios_back().status = results::test_status::failed;
+    std::for_each(results::scenarios_back().steps.begin(),
+                  results::scenarios_back().steps.end(),
+                  [&msg](results::step& step) { step.error_msg = msg; });
+  }
   else
   {
     const std::vector<results::step>& steps = results::scenarios_back().steps;
@@ -127,6 +138,7 @@ void update_scenario_status(std::string_view name, std::string_view file,
     }
   }
 
+  internal::get_runtime_options().reset_fail_scenario();
   results::test_results().add_scenario(results::scenarios_back().status);
   update_feature_status();
 }
@@ -134,6 +146,7 @@ void update_scenario_status(std::string_view name, std::string_view file,
 void update_step_status()
 {
   results::test_results().add_step(results::steps_back().status);
+  internal::get_runtime_options().reset_fail_step();
 }
 
 }  // namespace
@@ -230,8 +243,10 @@ class test_runner
     {
       run_step(step, skip);
     }
-
-    cuke::registry().run_hook_after(scenario.tags());
+    if (!internal::get_runtime_options().fail_scenario().is_set)
+    {
+      cuke::registry().run_hook_after(scenario.tags());
+    }
     update_scenario_status(scenario.name(), scenario.file(), scenario.line(),
                            skip);
     cuke::internal::reset_context();
@@ -256,7 +271,18 @@ class test_runner
 
   void run_step(const ast::step_node& step, bool skip) const
   {
-    if (skip_step() || skip)
+    // FIXME: these two calls to results::new_step are confusing.
+    // right now skip_step() relies on it which means i can not put it before
+    // the if statement. Will fix later
+    //
+    // Plus: skip_step and skip seem not very clear in the first place:
+    //   - skip_step() -> if a prev. step failed (and the continue on failure
+    //   flag wasn't set)
+    //   - skip -> the skip flag from the user (cuke::skip_scenario()),
+    //     these steps are as skipped in the report
+    //
+    if (skip_step() || skip ||
+        internal::get_runtime_options().fail_scenario().is_set)
     {
       results::new_step(step);
       results::steps_back().status = step.has_step_definition()
@@ -271,8 +297,19 @@ class test_runner
     {
       results::set_source_location(step.source_location_definition());
       cuke::registry().run_hook_before_step();
-      step.call();
-      cuke::registry().run_hook_after_step();
+      if (internal::get_runtime_options().fail_step().is_set)
+      {
+        results::steps_back().status = results::test_status::failed;
+        results::steps_back().error_msg =
+            internal::get_runtime_options().fail_step().msg;
+        log::error(internal::get_runtime_options().fail_step().msg,
+                   log::new_line);
+      }
+      else
+      {
+        step.call();
+        cuke::registry().run_hook_after_step();
+      }
     }
     else
     {