+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added 
+
+- Manually fail a scenario or step with `cuke::fail_scenario()`, `cuke::fail_scenario("Optional error message")`, `cuke::fail_step()` or `cuke::fail_step("Optional error message")` ([89](https://github.com/ThoSe1990/cwt-cucumber/issues/89))
+- Program option `--exclude-file` to exclude a single file from the test run ([89](https://github.com/ThoSe1990/cwt-cucumber/issues/89))
+
 ## [2.7] 2025-06-16
 
 ### Added